x80: weight the -g profiler by cycles instead of execution count
Each instruction execution counted as 1 hit regardless of cost, so a
23-cycle call and a 4-cycle inc hl looked identical in a profile - and
a single large ldir/lddr/cpir/cpdr, which can dominate a routine's
runtime, showed as just "1 hit". handle_state() now attributes each
address the actual cycle delta since the previously-recorded
instruction, using the running "cycles" total x80_emulate_impl already
tracks.

Two correctness issues came with it, both fixed:
 - x80_emulate_impl resets "cycles" to 0 on every call (it's invoked
   repeatedly with a fixed maxcycles from the main loop), so a raw
   diff across that boundary would wrap into a bogus multi-thousand-
   cycle spike attributed to a random address. Detected via the delta
   going negative (cycles only increases within a batch) and skipped
   instead of computed.
 - That skip alone silently dropped the last instruction of every
   batch, which - discovered while fixing the above - could be a
   large ldir/etc. costing up to ~1.38M cycles on its own. Added an
   explicit flush at the batch's exit point.
 - uint16_t cycles (in both z80_emulate and x80_emulate_impl) can't
   hold that ~1.38M-cycle worst case either; widened the whole chain
   (z80_emulate, x80_emulate_impl, x80_emulate, handle_state) to
   uint32_t.

Verified against a real profiling run: the CSV's cycle sum now
matches ntvcm's own reported total cycle count exactly, and an ldir
that previously showed "1 hit" now shows its true ~101K-cycle cost.

+++ x80.hxx
@@ -288,7 +288,7 @@ extern void x80_hard_exit( const char * pcerror, uint8_t arg1, uint8_t arg2 ); /
 
 // emulator API
 
-extern uint16_t x80_emulate( uint16_t maxcycles );             // execute up to about maxcycles
+extern uint32_t x80_emulate( uint16_t maxcycles );             // execute up to about maxcycles
 extern void x80_trace_instructions( bool trace );              // enable/disable tracing each instruction
 extern void x80_end_emulation();                               // stop the emulation
 extern void x80_trace_state( void );                           // trace the registers
